+++ controllers/passport/index.ts
@@ -24,7 +24,7 @@ const social_type = {
 }
 
 passport.serializeUser(function (user, done) {
-  console.log(user);
+  // console.log(user);
   done(null, user);
 })
