+++ controllers/user/update/username.ts
@@ -4,7 +4,7 @@ import { social_type, user } from '../../helper'
 
 const { User } = db;
 
-const func = async (req: Request , res: Response) => {
+export const updateUsername = async (req: Request , res: Response) => {
   if (req.user) {
     let user = <user>req.user
 
@@ -18,7 +18,7 @@ const func = async (req: Request , res: Response) => {
       await result.update({
         username: req.body.username
       });
-      res.status(200).json({username: req.body.username});
+      res.status(200).send();
 
     } else {
       res.status(404).send();
@@ -28,5 +28,3 @@ const func = async (req: Request , res: Response) => {
     res.status(404).send();
   }
 }
-
-export default func;
