+++ package.json
@@ -10,7 +10,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "node": "12.19.0",
     "@types/aws-sdk": "^2.7.0",
     "@types/multer-s3": "^2.7.8",
     "@types/node": "^14.11.8",
@@ -23,7 +22,9 @@
     "express-session": "^1.17.1",
     "multer": "^1.4.2",
     "multer-s3": "^2.9.0",
+    "multer-sharp-s3": "^0.2.1",
     "mysql2": "^2.2.5",
+    "node": "12.19.0",
     "nodemon": "^2.0.4",
     "passport": "^0.4.1",
     "passport-google-oauth": "^2.0.0",
