+++ routes/user-update.ts
@@ -1,9 +1,10 @@
 import * as express from 'express'
-import { updateProfileIMG, updateUsername, upload } from '../controllers/user/update';
+import { updateProfileIMG, updateUsername } from '../controllers/user/update';
+import { resizeUpload } from '../controllers/multer';
 
 const route = express.Router();
 
 route.post('/username', updateUsername);
-route.post('/profile-img', upload.single('img'), updateProfileIMG);
+route.post('/profile-img', resizeUpload.single('img'), updateProfileIMG);
 
 export default route;
