+++ tsconfig.json
@@ -2,6 +2,7 @@
     "compilerOptions": {
         "module": "CommonJS",
         "target": "es6",
+        "esModuleInterop": true
     },
     "types": [
         "node"
